This is the driver for the class graph, but the one done by miquel, I have to update it with the new things.
But I've got a problem, to use the functions like edgeExists(Edge e) I ask the user to enter parameters to create a new edge, as it will be a different object from the one in the actual graph (although it may have the same values, it is a phisically different object) the function may return false, and I don't know how to do that. I am assuming that what is compared are the objects, not the actual values, which I think it would be more useful in our case. In the case of the nodes, I have solved that problem using a map where every node entered is saved then when you reuse a node, it is picked from there so that it will be the same object.